refactor(pedagogy): extract apply tests to sibling file
`dialogue_manager/apply.rs` carried a single inline trailing test
module of pure-function apply_* unit tests. Move the whole module
verbatim into a sibling `apply/tests.rs` and replace the inline block
with `mod tests;`. The test module is a descendant, so `use super::*`
resolves to `apply` unchanged — no visibility edits, zero churn.

apply.rs 588 -> 186 lines (-402). All 203 primer-pedagogy lib tests
preserved; workspace 51 ok-blocks, clippy/fmt clean.
